requirements.txt
# Telegram Bot
aiogram>=3.0.0,<4.0.0

# Async SQLite
aiosqlite>=0.19.0

# Exchange API (Async)
ccxt>=4.2.0

# Environment & Config
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# WebSocket
websockets>=12.0

# HTTP Client
aiohttp>=3.9.0

# Utils
typing-extensions>=4.8.0
